Rearrange code to make coverage more accurate.
I really don't understand what's going on here. The second "return
_ZSS_MAP_SKIP" line was being reported as uncovered. It should have
been covered, though, by the call to .validate() (which uses
skip_index=False) on the good-extension-blocks file -- and in fact,
when I added a print statement here, it turns out it was covered. And
what's more, adding a print statement made it suddenly *appear*
covered in the coverage output as well.

I was suspecting some sort of optimization folding the two return
_ZSS_MAP_SKIP lines together... but I couldn't see any evidence of
this in the disassembly. Swapping the order of the two if's is perhaps
slightly clearer in any case, and it seems to give the right coverage
output, so, ok, that's what we'll do.